# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayStream Contracts API
  description: |
    PayStream smart contract interface for real-time salary streaming on Stellar (Soroban).
    All contract calls are Soroban XDR invocations. This document describes function
    signatures, argument types, and return values.
  version: 1.0.0
  contact:
    email: support@paystream.example
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://soroban-testnet.stellar.org
    description: Stellar Testnet
  - url: https://soroban.stellar.org
    description: Stellar Mainnet
...
paths:
  /initialize:
    post:
      tags: [Admin]
      summary: Initialize contract admin
      operationId: initialize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [admin]
              properties:
                admin:
                  type: string
                  description: Stellar account address of the admin
                  example: GABC1234...
      responses:
        '200':
          description: Contract initialized
        '400':
          $ref: '#/components/responses/AlreadyInitialized'
...
  /withdraw:
    post:
      tags: [Employee]
      summary: Withdraw claimable earnings
      operationId: withdraw
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employee, stream_id]
              properties:
                employee:
                  type: string
                  example: GXYZ9876...
                stream_id:
                  type: integer
                  format: uint64
                  example: 42
      responses:
        '200':
          description: Amount withdrawn
          content:
            application/json:
              schema:
                type: integer
                format: int128
        '400':
          $ref: '#/components/responses/NothingToWithdraw'
...
components:
...
  responses:
...
    AlreadyInitialized:
      description: Contract already initialized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NothingToWithdraw:
      description: No claimable balance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...